Moving mesh quality in its own class
